requirements.txt
pydub
google-generativeai
chromadb
python-dotenv
openai
langchain_openai
langchain_chroma
langchain_text_splitters